+++ docs/source/CHANGELOG.rst
@@ -4,8 +4,18 @@ Changelog
 Changelog for PyTornado. Version numbers try to follow `Semantic
 Versioning <https://semver.org/spec/v2.0.0.html>`__.
 
-[0.1.1] -- 2019-09-04
+[0.1.1] -- 2019-09-15
 ---------------------
+
+Changed
+~~~~~~~
+
+* Modified arguments of `Airfoil.plot()` function
+
+Added
+~~~~~
+
+* Added `save` flag to `Airfoil.plot()` function
 
 Fixed
 ~~~~~